xdp: Move shared code to a new shared/ toplevel directory
