+++ document-portal/meson.build
@@ -21,7 +21,6 @@ xdg_permission_store_sources = [
   'xdg-permission-store.c',
   xdp_utils_sources,
   db_sources,
-  sd_escape_sources,
   permission_store_built_sources,
 ]
 
@@ -64,7 +63,6 @@ xdg_document_portal_sources = [
   'document-portal-fuse.c',
   xdp_utils_sources,
   db_sources,
-  sd_escape_sources,
   document_portal_built_sources,
   permission_store_built_sources,
 ]
