+++ shared/sd-escape.c
