+++ shared/sd-escape.h
