+++ shared/xdp-app-info-flatpak-private.h
