+++ shared/xdp-app-info-flatpak.c
