+++ shared/xdp-app-info-registry.c
