+++ shared/xdp-app-info-registry.h
