+++ shared/xdp-app-info-snap-private.h
