+++ shared/xdp-app-info-snap.c
