+++ shared/xdp-usb-query.h
